move handler up
