+++ inngest/src/lib.rs
@@ -1,5 +1,6 @@
 pub mod event;
 pub mod function;
+pub mod handler;
 pub mod inngest;
 pub mod result;
 pub mod router;
